fix navbar con authcontext
